+++ ceiaun_bot/bot/conversations/__init__.py
@@ -1,0 +1,1 @@
+from .main import back_home, convert_course_handler, home_handler, request_course_handler, start_handler  # noqa
